Update and Implement new Code

+++ src/main/java/me/ry4nn00b/hortifruti/Model/SaleItemModel.java
@@ -4,12 +4,12 @@ public class SaleItemModel {
 
     //SaleItem Attribute's
     private String productId;
-    private Integer amount;
+    private Double amount;
     private Double unitPrice;
 
     public SaleItemModel() {}
 
-    public SaleItemModel(String productId, Integer amount, Double unitPrice) {
+    public SaleItemModel(String productId, Double amount, Double unitPrice) {
         this.productId = productId;
         this.amount = amount;
         this.unitPrice = unitPrice;
@@ -19,8 +19,8 @@ public class SaleItemModel {
     public String getProductId() { return productId; }
     public void setProductId(String productId) { this.productId = productId; }
 
-    public Integer getAmount() { return amount; }
-    public void setAmount(Integer amount) { this.amount = amount; }
+    public Double getAmount() { return amount; }
+    public void setAmount(Double amount) { this.amount = amount; }
 
     public Double getUnitPrice() { return unitPrice; }
     public void setUnitPrice(Double unitPrice) { this.unitPrice = unitPrice; }
